Restructure starlette client to conform with new project structure
