+++ authlib/integrations/starlette_client/__init__.py
@@ -1,0 +1,7 @@
+# flake8: noqa
+
+from .oauth_registry import OAuth
+from .remote_app import RemoteApp
+
+
+__all__ = ["OAuth", "RemoteApp"]
