+++ tests/starlette/test_oauth_client.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 from starlette.applications import Starlette as App
 from starlette.requests import Request
 from starlette.middleware.sessions import SessionMiddleware
-from authlib.starlette.client import OAuth
+from authlib.integrations.starlette_client import OAuth
 from ..client_base import mock_send_value, get_bearer_token
